+++ src/main/java/com/moing/backend/domain/auth/application/service/google/GoogleTokenUseCase.java
@@ -10,7 +10,7 @@ import java.util.Arrays;
 
 @Component
 @RequiredArgsConstructor
-public class GoogleTokenUserCase {
+public class GoogleTokenUseCase {
 
     @Value("${oauth2.google.appId}")
     private String appId;
