+++ src/main/java/com/moing/backend/domain/auth/application/service/google/utils/GoogleClient.java
@@ -1,6 +1,6 @@
 package com.moing.backend.domain.auth.application.service.google.utils;
 
-import com.moing.backend.global.util.FeignClientConfig;
+import com.moing.backend.global.utils.FeignClientConfig;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
